Add files via upload
Makes the process of gmail customisation easier by creating an interface where user can enter required keywords that they never want to miss, and also create a priority emails list. This process makes it easy and user-friendly to organise gmail, and import filters and implement this within 2 clicks on gmail.